requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Gemini Vision (replaces Google Cloud Vision)
google-generativeai==0.8.4

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# Fuzzy Matching
rapidfuzz==3.9.3

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.7.1
